Renamed TopBar to profilepagebanners and moved inside profile, same for sharepage
