+++ src/components/Profile/ProfilePage/ProfilePageBanners/NoWifiBar.tsx
